Create DualCameraQRRecognition
Utilized OpenCv to capture and examine QR codes to send an alert to the Arduino through Serial communication when the correct QR code had been found. Both cameras would be scanning for the correct QR code, then would send a signal to the Arduino controlling the Wheels to either turn left or right to get  to the door for the delivery.